Menu Navigation
- added navigation to all menu buttons
- made profile and lobby scenes
