+++ PlatformRacer/Assets/Scripts/CharacterCustomizer.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8ce511ef49a004341a40aeafd95b5f0f
