+++ PlatformRacer/Assets/Scripts/ExampleMenuManger.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9cff4e2fcd96eda47aab3d4aa21e9eb5
