+++ PlatformRacer/Assets/Scripts/MenuManager.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: 9cff4e2fcd96eda47aab3d4aa21e9eb5
+guid: 83c69cac91e2aaa4687a0851ad00a817
