+++ PlatformRacer/Assets/Scripts/SettingsManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b5c2d7fb4fe6b244f94d6fb7729a0821
