+++ PlatformRacer/Assets/Scripts/SliderTextUpdater.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cd3a050efde9132449fe56c23bad2445
