+++ PlatformRacer/PlatformRacer.slnx
@@ -1,0 +1,3 @@
+﻿<Solution>
+  <Project Path="Assembly-CSharp.csproj" />
+</Solution>
