+++ PlatformRacer/ProjectSettings/EditorBuildSettings.asset
@@ -8,6 +8,15 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/SampleScene.unity
     guid: 8c9cfa26abfee488c85f1582747f6a02
+  - enabled: 1
+    path: Assets/Scenes/LobbyScene.unity
+    guid: eb240d93100d2084694270ccfebe1adc
+  - enabled: 1
+    path: Assets/Scenes/ProfileScene.unity
+    guid: e75c34b1cdc66f74cb5c6941439e66b0
+  - enabled: 1
+    path: Assets/Scenes/MenuScene.unity
+    guid: 152293dc59bfbdc40b611e03eece68a6
   m_configObjects:
     com.unity.input.settings.actions: {fileID: -944628639613478452, guid: 2bcd2660ca9b64942af0de543d8d7100, type: 3}
   m_UseUCBPForAssetBundles: 0
